Few tweaks

+++ Tests/testsLib.cpp
@@ -28,10 +28,11 @@ int main(int argc, char **argv) {
   int index = 0;
 
   while (pmap[index].s != 0) {
-    fprintf(stderr, "%p-%p %" PRIu64 " - huge ? %d (%ld)\n",
+    fprintf(stderr, "%p-%p %" PRIu64 " - huge ? %d (%" PRIu64 ")\n",
             reinterpret_cast<void *>(pmap[index].s),
-            reinterpret_cast<void *>(pmap[index].e), pmap[index].sz,
-            pmap[index].hgmp, pmap[index].f);
+            reinterpret_cast<void *>(pmap[index].e),
+            static_cast<int64_t>(pmap[index].sz), pmap[index].hgmp,
+            pmap[index].f);
     ++index;
   }
